+++ packages/smarthr-ui/src/components/Dialog/ControlledActionDialog/ActionDialog.test.tsx
@@ -16,7 +16,7 @@ describe('ControlledActionDialog', () => {
         <ControlledActionDialog
           isOpen={isOpen}
           heading="ControlledActionDialog"
-          actionText="保存"
+          actionButton="保存"
           onClickClose={() => setIsOpen(false)}
           onClickAction={(_, { close }) => {
             close()