+++ packages/smarthr-ui/src/components/Dialog/ControlledFormDialog/ControlledFormDialog.tsx
@@ -2,6 +2,7 @@
 
 import { type ComponentProps, type FC, type FormEvent, type ReactNode, useCallback } from 'react'
 
+import { useObjectAttributes } from '../../../hooks/useObjectAttributes'
 import { DialogContentInner } from '../DialogContentInner'
 import { useDialogPortal } from '../useDialogPortal'
 import { useObjectHeading } from '../useObjectHeading'
@@ -16,34 +17,36 @@ import type { DialogProps } from '../types'
 
 type ObjectHeadingType = Omit<FormDialogContentInnerProps['heading'], 'id'>
 type HeadingType = ReactNode | ObjectHeadingType
+type ObjectActionButtonType = FormDialogContentInnerProps['actionButton']
+type ObjectCloseButtonType = FormDialogContentInnerProps['closeButton']
 
-type AbstractProps = Omit<FormDialogContentInnerProps, 'heading'> &
+type AbstractProps = Omit<FormDialogContentInnerProps, 'heading' | 'actionButton' | 'closeButton'> &
   DialogProps & {
     heading: HeadingType
+    actionButton: ReactNode | ObjectActionButtonType
+    closeButton?: ReactNode | ObjectCloseButtonType
   }
 type Props = AbstractProps & Omit<ComponentProps<'div'>, keyof AbstractProps>
 
 const headingObjectConverter = (text: ReactNode) => ({
   text,
 })
+const buttonObjectConverter = (text: ReactNode) => ({ text })
 
 export const ControlledFormDialog: FC<Props> = ({
   children,
   heading: orgHeading,
   contentBgColor,
   contentPadding,
-  actionText,
-  actionTheme,
+  actionButton: orgActionButton,
   onSubmit,
   onClickClose,
   onPressEscape = onClickClose,
   responseStatus,
-  actionDisabled,
-  closeDisabled,
+  closeButton: orgCloseButton,
   subActionArea,
   className,
   portalParent,
-  decorators,
   id,
   isOpen,
   ...rest
@@ -52,6 +55,14 @@ export const ControlledFormDialog: FC<Props> = ({
   const heading = useObjectHeading<HeadingType, ObjectHeadingType>(
     orgHeading,
     headingObjectConverter,
+  )
+  const actionButton = useObjectAttributes<
+    ReactNode | ObjectActionButtonType,
+    ObjectActionButtonType
+  >(orgActionButton, buttonObjectConverter)
+  const closeButton = useObjectAttributes<ReactNode | ObjectCloseButtonType, ObjectCloseButtonType>(
+    orgCloseButton,
+    buttonObjectConverter,
   )
 
   const actualOnClickClose = useCallback(() => {
@@ -75,21 +86,18 @@ export const ControlledFormDialog: FC<Props> = ({
       isOpen={isOpen}
       ariaLabelledby={heading.id}
       className={className}
-      onPressEscape={closeDisabled ? undefined : onPressEscape}
+      onPressEscape={closeButton.disabled ? undefined : onPressEscape}
     >
       <FormDialogContentInner
         heading={heading}
         contentBgColor={contentBgColor}
         contentPadding={contentPadding}
-        actionText={actionText}
-        actionTheme={actionTheme}
-        actionDisabled={actionDisabled}
-        closeDisabled={closeDisabled}
+        actionButton={actionButton}
+        closeButton={closeButton}
         subActionArea={subActionArea}
         onClickClose={actualOnClickClose}
         onSubmit={onDelegateSubmit}
         responseStatus={responseStatus}
-        decorators={decorators}
       >
         {children}
       </FormDialogContentInner>